Move CoreConfigTest and MockKubeConfig to root core test package
